+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 * Add support for querying on RealmSets containing objects with `RealmSet.query(...)`.  (Issue [#1037](https://github.com/realm/realm-kotlin/issues/1258))
 * [Sync] Add support for setting App Services connection identifiers through `AppConfiguration.appName` and `AppConfiguration.appVersion`, making it easier to identify connections in the server logs.
 . (Issue (#407)[https://github.com/realm/realm-kotlin/issues/407])
- 
+* Add Realm datatypes serialization support for `Kserializer`. Serializers can be found in `io.realm.kotlin.serializers`. (Issue [#1283](https://github.com/realm/realm-kotlin/pull/1283))
+
 ### Fixed
 * None.